Split intial drawing functions into seperate file
